feat(WinLotto): 당첨 클래스 구현

+++ src/main/java/lotto/commons/WinLotto.java
@@ -22,7 +22,7 @@ public class WinLotto {
         return winHistory;
     }
 
-    public int getWinning() {
+    public long getWinning() {
         return winning;
     }
